Pass app state to commands
We rename the first command param to app to note that the app state is
being passed to the command, and then we redo the object being passed
into commands to make it more clear what's passed. We'll use this setup
to introduce a SQLite3 connection object to perform database operations
without establishing a connection per request.