Correctly handle automatic date parsing
Previously, I made a very crude implementation to calculate peoples
dates for day am/pm to add to the correct record/column in the price
lookup. By adding in the TurnupDateCalculator & corresponding tests, we
can correctly handle automatic & manual date parsing for users.

This update will only handle timezones correctly, if a date with a
timezone is passed in correctly--which currently, that's not the case.
We're still relying on server timezone.