Add timezone command
Discord only sends messages with a UTC datetime, and does not include a
client local datetime. Because of this, we need to introduce a timezone
command that'll allow us to set user timezones. We'll use this to query
the userMeta table to pull the timezone to convert the discord message
time into the users local time so that their dates can automatically be
calculated.