Rework commands
Commands didn't provide a good a good way to provide detailed help
information as well as were all implemented in a single file. This
update creates an actual Command class which all commands inherit, and
allow for the bot to have a robust help system. Also, this update
improves code duplication problems, and makes it easier to implement
commands.