TurnipService: Fix timezone conversion bug
In the case where there's no userMeta for a user, the
TurnipService::getTurnipDateCalculator() would attempt to use the users
timezone which would be null and cause a "TypeError: Cannot read
property 'timezone' of null" error. We fix this by guarding the timezone
conversions and to only do it if there's userMeta.